requirements.txt
docker-py >= 1.6.0

